requirements.txt
# Core data analysis
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
scipy==1.12.0
statsmodels==0.14.1
scikit-learn==1.4.1.post1

# Financial data
yfinance==0.2.37
pandas-datareader==0.10.0

# Jupyter notebook support
jupyter==1.0.0
ipykernel==6.29.0
ipywidgets==8.1.2
nbformat==5.10.2

# Data processing
openpyxl==3.1.2
xlrd==2.0.1

# Dev tools
pytest==8.0.0
black==24.3.0
python-dotenv==1.0.1
telethon==1.32.0 
